// src/type/type.tsx
export interface ICharacter {
    id: number;
    name: string;
    status: string;
    species: string;
    type: string;
    gender: string;
    origin: Location;
    location: Location;
    image: string;
    episode: string[];
    url: string;
    created: string;
  }
  

  export interface IEpisode {
    id: number
    name: string
    type: string
    dimension: string
    residents: string[]
    url: string
    created: string
    air_date:string
    episode:string
  }
  